LayoutTests/storage/indexeddb/modern/opendatabase-after-storage-crash-expected.txt

PASS localStorage[storageKey] is "Made it"
Successfully opened database with version 3.
